将安装脚本中why的路径的改为脚本路径

+++ install.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-rm -rf $PREFIX/bin/why
-cp why $PREFIX/bin
+if [[ -f $PREFIX/bin/why]] ; then
+  rm -rf $PREFIX/bin/why
+fi
+cp $(cd `dirname $0`; pwd)/why $PREFIX/bin/
 chmod +x $PREFIX/bin/why
 echo "安装已完成"
